refactor: move non parser types and vars out of `stylable-value-parsers`
- `MappedStates` to `features/types` and expose internaly
- `ReportWarning` to `helpers/value`
- deprecate `stKeys`, `stValues`, `stValuesMap`, `STYLABLE_NAMED_MATCHER`, `mixinDeclRegExp`, `animationPropRegExp`, `STYLABLE_VALUE_MATCHER`, `TypedClass`
