+++ packages/core/src/deprecated/leftovers.ts
@@ -1,0 +1,8 @@
+import type { MappedStates } from '../features';
+
+/**@deprecated */
+export interface TypedClass {
+    '-st-root'?: boolean;
+    '-st-states'?: string[] | MappedStates;
+    '-st-extends'?: string;
+}
