+++ packages/core/src/features/types.ts
@@ -1,10 +1,12 @@
 import type { ImportSymbol } from './st-import';
 import type { ClassSymbol } from './css-class';
 import type { ElementSymbol } from './css-type';
-import type { MappedStates } from '../stylable-value-parsers';
+import type { StateParsedValue } from '../types';
 import type { SelectorNode } from '@tokey/css-selector-parser';
 
-// ToDo: distribute types to features
+export interface MappedStates {
+    [s: string]: StateParsedValue | string | null;
+}
 
 export interface StylableDirectives {
     '-st-root'?: boolean;
