+++ packages/core/src/index-internal.ts
@@ -1,2 +1,3 @@
 export { safeParse } from './parser';
 export { process } from './stylable-processor';
+export type { MappedStates } from './features';
