+++ packages/core/src/pseudo-states.ts
@@ -14,9 +14,8 @@ import type { PseudoClass } from '@tokey/css-selector-parser';
 import { StateResult, systemValidators } from './state-validators';
 import type { StylableMeta } from './stylable-meta';
 import type { StylableResolver } from './stylable-resolver';
-import type { MappedStates } from './stylable-value-parsers';
 import type { ParsedValue, StateParsedValue } from './types';
-import { CSSClass } from './features';
+import { CSSClass, MappedStates } from './features';
 import { stripQuotation } from './helpers/string';
 import { reservedFunctionalPseudoClasses } from './native-reserved-lists';
 import cssesc from 'cssesc';
