+++ packages/module-utils/src/generate-dts.ts
@@ -1,11 +1,11 @@
 import type {
     ClassSymbol,
-    MappedStates,
     StateParsedValue,
     StylableMeta,
     StylableResults,
     StylableSymbol,
 } from '@stylable/core';
+import type { MappedStates } from '@stylable/core/dist/index-internal';
 
 export const SPACING = ' '.repeat(4);
 const asString = (v: string) => JSON.stringify(v);
